+++ src/extendAt.js
@@ -6,19 +6,19 @@ import getAt from './getAt';
 import replaceAt from './replaceAt';
 import extend from './extend';
 
-const extendAt = (src, path, ext) => {
-  const pathArray = toPath(path);
+const extendAt = (src, pathToObject, extensionOrCreator) => {
+  const pathArray = toPath(pathToObject);
   // console.log({ path, pathArray });
   const valueAt = getAt(src, pathArray);
   // console.log('valueAt === src:', valueAt === src, pathArray);
-  const targetAt = extend(valueAt, ext);
+  const targetAt = extend(valueAt, extensionOrCreator);
   // console.log({ valueAt, targetAt, pathArray });
   if (targetAt === valueAt) {
     /** Not changed at path */
     return src;
   } else {
     /** Changed at path |=> Immutable replace at path */
-    return replaceAt(src, toPath(path), targetAt);
+    return replaceAt(src, toPath(pathToObject), targetAt);
   }
 };